+++ src/model/App.tsx
@@ -191,10 +191,12 @@ export class App {
   runTestModeBytes0To255() {
     console.log('runTestMode2() called.');
     this.settings.dataProcessing.visibleData.fields.ansiEscapeCodeParsingEnabled.value = false;
+    this.settings.dataProcessing.visibleData.fields.charSizePx.value = '30';
     this.settings.dataProcessing.applyChanges();
     let testCharIdx = 0;
     setInterval(() => {
-      this.parseRxData(Uint8Array.from([ testCharIdx ]));
+      // this.parseRxData(Uint8Array.from([ testCharIdx ]));
+      this.parseRxData(Uint8Array.from([ 0x08 ]));
       testCharIdx += 1;
       if (testCharIdx === 256) {
         testCharIdx = 0;